+++ src/main/java/com/sparta/zipsa/controller/MatchBoardController.java
@@ -3,9 +3,8 @@ package com.sparta.zipsa.controller;
 import com.sparta.zipsa.dto.MatchBoardRequestDto;
 import com.sparta.zipsa.dto.MatchBoardResponseDto;
 import com.sparta.zipsa.entity.MatchBoard;
-import com.sparta.zipsa.exception.MatchException;
 import com.sparta.zipsa.security.UserDetailsImpl;
-import com.sparta.zipsa.service.board.MatchBoardService;
+import com.sparta.zipsa.service.matchBoard.MatchBoardService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
